+++ src/lib/constants.ts
@@ -1,1 +1,6 @@
-export const availableMaps = ['world-1', 'world-2', 'world-3', 'nuts-1', 'nuts-2', 'nuts-3'];
+export const availableMaps = [
+	{ label: 'Countries', value: 'world-3' },
+	{ label: 'NUTS 1 (only Europe)', value: 'nuts-1' },
+	{ label: 'NUTS 2 (only Europe)', value: 'nuts-2' },
+	{ label: 'NUTS 3 (only Europe)', value: 'nuts-3' }
+];