+++ src/routes/transition/capacity/+page.svelte
@@ -29,7 +29,6 @@
 	let aggregation_options = $state(['technology', 'node']);
 	const normalisation_options = ['not_normalized', 'normalized'];
 	const storage_type_options = ['energy', 'power'];
-	let units: { [carrier: string]: string } = $state({});
 	let selected_variable: string | null = $state('capacity');
 	let selected_carrier: string | null = $state(null);
 	let selected_storage_type = $state('energy');
@@ -46,6 +45,12 @@
 
 	let datasets: any[] = $state([]);
 	let labels: string[] = $state([]);
+
+	let units: { [carrier: string]: string } = $state({});
+	let unit: string = $derived.by(() => {
+		const capacity_type = selected_technology_type == 'storage' ? selected_storage_type : 'power';
+		return units[technologies[0] + '_' + capacity_type] || '';
+	});
 
 	let plot_config: ChartConfiguration<'bar'> = $derived({
 		type: 'bar',
@@ -64,7 +69,7 @@
 					stacked: true,
 					title: {
 						display: true,
-						text: `${selected_variable} [${technologies.length > 0 ? units[technologies[0]] : ''}]`
+						text: `${selected_variable} [${unit}]`
 					}
 				}
 			}
@@ -91,7 +96,7 @@
 		data = null;
 		await tick();
 
-		if (selected_variable === null) {
+		if (selected_variable === null || selected_solution === null) {
 			fetching = false;
 			return;
 		}
@@ -107,7 +112,9 @@
 		data = fetched.data;
 
 		if (fetched.unit?.data) {
-			units = Object.fromEntries(fetched.unit.data.map((u) => [u.technology, u[0] || u.units]));
+			units = Object.fromEntries(
+				fetched.unit.data.map((u) => [u.technology + '_' + u.capacity_type, u[0] || u.units])
+			);
 		}
 
 		fetching = false;
@@ -195,9 +202,13 @@
 	 * This function updates the available locations for the current variable selection.
 	 */
 	function update_locations() {
+		if (data === null) {
+			return;
+		}
+
 		locations = [];
 
-		data!.data.forEach((element) => {
+		data.data.forEach((element) => {
 			let current_technology = element.technology;
 			let current_carrier = selected_solution!.detail.reference_carrier[current_technology];